backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
python-multipart
openpyxl
numpy
python-dotenv
orjson
httpx
pyarrow